% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust}
\alias{adjust}
\title{Correct carcass to nearest visible patch.}
\usage{
adjust(pto, ring)
}
\arguments{
\item{pto}{a vector of found carcasses.}

\item{ring}{a shapefile describing the visibility for each turbine.}
}
\value{
a sf object.
}
\description{
The \code{adjust} function will adjust the carcass position when its position
is over a not searched or not visible patch. Misplaced carcasses can happens due to gps
bad fix. Correcting positions will be relevant to correctly estimate DWP.
}
\details{
ensure that ring is a vector with the columns ag, dist and vis.
}
\examples{
# not run
ring <- viring(x, d)
adjust(pto, ring)
}
\author{
Paulo E. Cardoso
}
