% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwp.R
\name{dwp}
\alias{dwp}
\title{Concentric non-overpaping rings.}
\usage{
dwp(vr, pt)
}
\arguments{
\item{vr}{a shapefile describing the visibility for each turbine at each ring. Ideally should
be obtained with \code{virings}.}

\item{pt}{a shapefile with carcass positions with an unique ID for each one.}
}
\value{
a data.frame with dwp for each turbine at each ring distance.
}
\description{
The \code{dwp} function will create the table with Density Weighted Probability
required to run  \code{genest}.
}
\details{
Ensure that pt is a vector with the group you are interested in.
}
\examples{
# not run
ags <- st_sfc(st_point(c(1,2)), st_point(c(200,200)))
ags <- st_sf(geometry = ags) \%>\%
st_set_crs(3763) \%>\%
mutate(ag = 1:2)

# Distaces for concentric rings
dist = units::set_units(c(10, 20, 30, 40, 50, 100), m)

rings <- vrings(x = ags, d = dist)

# Carcass distritubion
logs <- st_sample(st_buffer(ags, 50), 10, type = "random", exact = TRUE) \%>\%
  st_sf(.)
logs$idu <- 1:5

ggplot() +
  geom_sf(aes(fill = as.numeric(dist), colour = factor(ag)),
          size = 1.1,
          data = rings) +
  geom_sf(data = logs)

# obtain dwp data.frame
dfdwp <- dwp(vr = rings, pt = logs)

}
\author{
Paulo E. Cardoso
}
