% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_carcass.R
\docType{data}
\name{pto_carcass}
\alias{pto_carcass}
\title{Shapefile with fake carcasses locations}
\format{A shapefile with 1 column and 10 lines
\describe{
  \item{ag}{turbine}
  \item{especie}{species id}
  \item{grupo}{Bird or Bat}
  \item{tamanho}{group size}
}}
\source{
Candeeiros WF monitoring
}
\usage{
pto_carcass
}
\description{
Shapefile with fake carcasses locations
}
\examples{
\dontrun{
summary(pto_carcass)
}
}
\author{
Paulo E. Cardoso
}
\keyword{datasets}
