% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visib.R
\docType{data}
\name{visib}
\alias{visib}
\title{Shapefile with fake visibility map.}
\format{}
\source{
QGIS
}
\usage{
visib
}
\description{
Shapefile with fake visibility map.
}
\examples{
\dontrun{
data(visib)
summary(visib)
}
}
\author{
Paulo E. Cardoso
}
\keyword{datasets}
