% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwp.R
\name{dwp}
\alias{dwp}
\title{Concentric non-overpaping rings.}
\usage{
dwp(vr, pt)
}
\arguments{
\item{vr}{a shapefile describing the visibility for each turbine.}

\item{pt}{a vector of distances to build concentric rings aroung each mapped area.}
}
\value{
a sf object.
}
\description{
The \code{virings} function will create a new layer crossing the visibility
mapped with concentric and non-overpaping rings for each turbine. Centroids are obtained from
the mapped area that will be ideally be built using the \code{rings}
}
\details{
ensure that d is a vector with distances, regular or not. Your visibility layer must
have the columns [ag] and [vis].
}
\author{
Paulo E. Cardoso
}
