% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virings.R
\name{viring}
\alias{viring}
\title{crossing  visibility with concentric rings.}
\usage{
viring(x, d)
}
\arguments{
\item{x}{a shapefile describing the visibility for each turbine.}

\item{d}{a vector of distances to build concentric rings aroung each mapped area.}
}
\value{
a sf object.
}
\description{
The \code{virings} function will create a new layer crossing the visibility
mapped with concentric and non-overpaping rings for each turbine. Centroids are obtained from
the mapped area that will be ideally be built using the \code{rings}
}
\details{
ensure that d is a vector with distances, regular or not. Your visibility layer must
have the columns ag and vis.
}
\examples{
# not run
rings <- viring(x, d)

ggplot() +
  geom_sf(aes(fill = as.numeric(area), colour = factor(ag)),
         size = .5,
          data = rings) +
  geom_sf(data = logs)

dfdwp <- dwp(vr = rings, pt = logs)
}
\author{
Paulo E. Cardoso
}
