% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virings.R
\name{viring}
\alias{viring}
\title{cross  visibility map with concentric rings.}
\usage{
viring(x, d)
}
\arguments{
\item{x}{a shapefile describing the visibility mapped for each turbine.}

\item{d}{a vector of distances to build concentric rings aroung each mapped area.}
}
\value{
a sf object.
}
\description{
The \code{virings} function will create a new layer resulting from crossing the
mapped visibility with concentric-and-non-overpaping rings for each mapped area around
turbines. Centroids are obtained from the visibility map.
Visibility map can be built from buffer areas resulting from \code{sf::st_buffer} function.
}
\details{
Ensure that d is a vector with distances, single, regularly spaced or not.
 Your visibility layer must have at least two columns: ag and visib.
 Check and remember to use a single projection for all layers in your project.
 Do not use geographic coordinate system.
}
\examples{
\dontrun{
# Distaces for rings
dist = c(10, 20, 30, 40, 50, 100)
# crossing visibility map with rings
rings <- viring(x = visib, d = dist)
# Add some randomly displaced carcasses
logs <- st_sample(st_buffer(ags, 50), 10, type = "random",
                  exact = TRUE) \%>\%
 st_sf(.)
logs$idu <- 1:5
# Plot it
ggplot() +
  geom_sf(aes(fill = as.numeric(area), colour = factor(ag)),
          size = .5,
          data = rings) +
  geom_sf(data = logs)

dfdwp <- dwp(vr = rings, pt = logs)
}
}
\author{
Paulo E. Cardoso
}
